Moved shared types to shared module common_utils and reorganised function exports in modules
Renamed FormFields to SubmittedOrderData and moved them into global_types.rs
