+++ web_server_with_database/src/database_handler/mod.rs
@@ -1,8 +1,5 @@
-mod database_handler;
-pub use database_handler::add_evaluation_to_db;
-pub use database_handler::add_form_submission_to_db;
-pub use database_handler::get_pending_order;
-pub use database_handler::initialize_db;
-pub use database_handler::read_orders_from_db;
-pub use database_handler::remove_order_from_db;
-pub use database_handler::FormFields;
+// PUBLIC MODULES
+pub mod database_handler;
+pub use database_handler::*;
+
+// PRIVATE MODULES
