+++ web_server_with_database/src/orca_slicer_interface/mod.rs
@@ -1,5 +1,6 @@
-mod orca_slicer_interface;
-pub use orca_slicer_interface::get_orca_slicer_evaluation;
-pub use orca_slicer_interface::initialize_orca_slicer_if;
-pub use orca_slicer_interface::EvaluationResult;
+// PUBLIC MODULES
+pub mod orca_slicer_interface;
+pub use orca_slicer_interface::*;
+
+// PRIVATE MODULES
 mod orca_slicer_cli;
